Added switch to load config from file
Now we can define wanted number of Priors inside configuration python file.

Regression and classification headers of the model will be created according to values inside configuration.

Use num_priors = [6,6,6,6,6,6] to define the number of default priors
per layer and
channels_priors = [512, 1024, 512, 256, 256, 256] to set the input
channels for the regression and classification headers